Splitted up IMap into IOriginalMap and IPlayableMap

+++ Rulesets/Maps/Timing/BreakPoint.cs
@@ -5,23 +5,23 @@ namespace PBGame.Rulesets.Maps.Timing
 		/// <summary>
 		/// Minimum number of milliseconds which a break point can be valid from.
 		/// </summary>
-		public const double MinBreakDuration = 650;
+		public const float MinBreakDuration = 650;
 
 		/// <summary>
 		/// The starting time of this break point
 		/// </summary>
-		public double StartTime { get; set; }
+		public float StartTime { get; set; }
 
 		/// <summary>
 		/// The ending time of this break point.
 		/// </summary>
-		public double EndTime { get; set; }
+		public float EndTime { get; set; }
 
 
 		/// <summary>
 		/// Returns the duration of this break point.
 		/// </summary>
-		public double Duration => EndTime - StartTime;
+		public float Duration => EndTime - StartTime;
 
         /// <summary>
         /// Returns whether this break point is valid and can provide breaks.
@@ -31,6 +31,6 @@ namespace PBGame.Rulesets.Maps.Timing
         /// <summary>
         /// Returns whether a valid breakpoint can be created with specified time values.
         /// </summary>
-        public static bool CanBeValid(double start, double end) => (end - start) > MinBreakDuration;
+        public static bool CanBeValid(float start, float end) => (end - start) > MinBreakDuration;
 	}
 }
